Refactored code into separate python files for maintainability, added some comments, and tested all methods on DBR 10.4 LTS
