+++ labelspark/constants.py
@@ -1,0 +1,8 @@
+LABELBOX_DEFAULT_TYPE_DICTIONARY = {
+    'Agreement': 'integer',
+    'Benchmark Agreement': 'integer',
+    'Created At': 'timestamp',
+    'Updated At': 'timestamp',
+    'Has Open Issues': 'integer',
+    'Seconds to Label': 'float',
+}
