+++ labelspark/is_json.py
@@ -1,0 +1,7 @@
+import json
+def is_json(myjson):
+    try:
+        json_object = json.loads(myjson)
+    except Exception as e:
+        return False
+    return True
